requirements/development.txt
-r base.txt
pytest-django>=4.8
pytest-cov>=5.0
coverage>=7.5
